+++ src/orquestra/sdk/_base/_driver/_client.py
@@ -12,7 +12,7 @@ import io
 import json
 import zlib
 from tarfile import TarFile
-from typing import Generic, List, Mapping, Optional, TypeVar
+from typing import Generic, List, Mapping, Optional, Tuple, TypeVar, Union
 from urllib.parse import urljoin
 
 import pydantic
@@ -21,7 +21,7 @@ from requests import codes
 
 from orquestra.sdk._ray._ray_logs import WFLog
 from orquestra.sdk.schema.ir import WorkflowDef
-from orquestra.sdk.schema.responses import WorkflowResult
+from orquestra.sdk.schema.responses import ComputeEngineWorkflowResult, WorkflowResult
 from orquestra.sdk.schema.workflow_run import (
     ProjectRef,
     WorkflowRun,
@@ -560,7 +560,7 @@ class DriverClient:
 
     def get_workflow_run_result(
         self, result_id: _models.WorkflowRunResultID
-    ) -> WorkflowResult:
+    ) -> Union[WorkflowResult, ComputeEngineWorkflowResult]:
         """
         Gets workflow run results from the workflow driver
 
@@ -583,11 +583,28 @@ class DriverClient:
 
         _handle_common_errors(resp)
 
-        # Bug with mypy and Pydantic:
-        #   Unions cannot be passed to parse_obj_as: pydantic/pydantic#1847
-        return pydantic.parse_obj_as(
-            WorkflowResult, resp.json()  # type: ignore[arg-type]
-        )
+        # To ensure the correct ordering of results, we serialize the results on CE as:
+        # {
+        #   "results": [
+        #       (JSONResult | PickleResult).json(),
+        #       (JSONResult | PickleResult).json(),
+        #       ...
+        #   ]
+        # } aka a ComputeEngineWorkflowResult.json()
+        # For older workflows, we respond with:
+        # (JSONResult | PickleResult).json()
+
+        json_response = resp.json()
+        try:
+            # Try an older response
+            # Bug with mypy and Pydantic:
+            #   Unions cannot be passed to parse_obj_as: pydantic/pydantic#1847
+            return pydantic.parse_obj_as(
+                WorkflowResult, json_response  # type: ignore[arg-type]
+            )
+        except pydantic.ValidationError:
+            # If we fail, try parsing each part of a list separately
+            return ComputeEngineWorkflowResult.parse_obj(json_response)
 
     # --- Workflow Logs ---