+++ tests/cli/dorq/test_repos.py
@@ -419,6 +419,9 @@ class TestWorkflowRunRepo:
                                 function_name="task_in_another_module",
                             ),
                             parameters=[],
+                            output_metadata=ir.TaskOutputMetadata(
+                                is_subscriptable=False, n_outputs=1
+                            ),
                             source_import_id="imp1",
                         ),
                         ir.TaskDef(
@@ -428,6 +431,9 @@ class TestWorkflowRunRepo:
                                 function_name="task_in_another_file",
                             ),
                             parameters=[],
+                            output_metadata=ir.TaskOutputMetadata(
+                                is_subscriptable=False, n_outputs=1
+                            ),
                             source_import_id="imp1",
                         ),
                         ir.TaskDef(
@@ -436,6 +442,9 @@ class TestWorkflowRunRepo:
                                 function_name="inlined_task", encoded_function=[]
                             ),
                             parameters=[],
+                            output_metadata=ir.TaskOutputMetadata(
+                                is_subscriptable=False, n_outputs=1
+                            ),
                             source_import_id="imp1",
                         ),
                     ]
@@ -470,6 +479,9 @@ class TestWorkflowRunRepo:
                                 module="tasks1", function_name=fn_name
                             ),
                             parameters=[],
+                            output_metadata=ir.TaskOutputMetadata(
+                                is_subscriptable=False, n_outputs=1
+                            ),
                             source_import_id="imp1",
                         ),
                         ir.TaskDef(
@@ -478,6 +490,9 @@ class TestWorkflowRunRepo:
                                 module="tasks2", function_name=fn_name
                             ),
                             parameters=[],
+                            output_metadata=ir.TaskOutputMetadata(
+                                is_subscriptable=False, n_outputs=1
+                            ),
                             source_import_id="imp1",
                         ),
                     ]
@@ -539,30 +554,91 @@ class TestWorkflowRunRepo:
                     _ = repo.get_task_inv_ids(wf_run_id, config_name, task_fn_name)
 
         class TestGetTaskOutputs:
-            @staticmethod
-            def test_passing_data(monkeypatch):
-                run_id = "wf.1"
-                inv_id = "inv1"
-                config_name = "<config sentinel>"
+            class TestHappyPath:
+                @staticmethod
+                def _make_wf_run(
+                    monkeypatch,
+                    inv_id: str,
+                    fake_task_run_outputs,
+                    task_meta: ir.TaskOutputMetadata,
+                ) -> sdk.WorkflowRun:
+                    task_run = create_autospec(sdk.TaskRun)
+                    task_run.task_invocation_id = inv_id
+                    task_run.get_outputs.return_value = fake_task_run_outputs
 
-                wf_run = Mock()
-                fake_outputs = ("<output sentinel 0>", "<output sentinel 1>")
-                wf_run.get_artifacts.return_value = {
-                    inv_id: fake_outputs,
-                }
+                    task_def = create_autospec(ir.TaskDef)
+                    task_def_id = "task_def_1"
+                    task_def.output_metadata = task_meta
 
-                by_id = Mock(return_value=wf_run)
-                monkeypatch.setattr(sdk.WorkflowRun, "by_id", by_id)
+                    inv = create_autospec(ir.TaskInvocation)
+                    inv.task_id = task_def_id
 
-                repo = _repos.WorkflowRunRepo()
+                    wf_def = create_autospec(ir.WorkflowDef)
+                    wf_def.task_invocations = {inv_id: inv}
+                    wf_def.tasks = {task_def_id: task_def}
 
-                # When
-                outputs = repo.get_task_outputs(
-                    wf_run_id=run_id, task_inv_id=inv_id, config_name=config_name
-                )
+                    wf_run = create_autospec(sdk.WorkflowRun)
+                    wf_run.get_tasks.return_value = {task_run}
+                    wf_run.get_status_model().workflow_def = wf_def
 
-                # Then
-                assert outputs == fake_outputs
+                    by_id = Mock(return_value=wf_run)
+                    monkeypatch.setattr(sdk.WorkflowRun, "by_id", by_id)
+
+                    return wf_run
+
+                def test_single_output(self, monkeypatch):
+                    # Given
+                    run_id = "wf.1"
+                    inv_id = "inv1"
+                    config_name = "<config sentinel>"
+                    fake_output = "<task output sentinel>"
+
+                    self._make_wf_run(
+                        monkeypatch,
+                        inv_id=inv_id,
+                        fake_task_run_outputs=fake_output,
+                        task_meta=ir.TaskOutputMetadata(
+                            n_outputs=1,
+                            is_subscriptable=False,
+                        ),
+                    )
+
+                    repo = _repos.WorkflowRunRepo()
+
+                    # When
+                    outputs = repo.get_task_outputs(
+                        wf_run_id=run_id, task_inv_id=inv_id, config_name=config_name
+                    )
+
+                    # Then
+                    assert outputs == (fake_output,)
+
+                def test_multiple_outputs(self, monkeypatch):
+                    # Given
+                    run_id = "wf.1"
+                    inv_id = "inv1"
+                    config_name = "<config sentinel>"
+                    fake_outputs = ("<out1>", "<out2>")
+
+                    self._make_wf_run(
+                        monkeypatch,
+                        inv_id=inv_id,
+                        fake_task_run_outputs=fake_outputs,
+                        task_meta=ir.TaskOutputMetadata(
+                            n_outputs=2,
+                            is_subscriptable=True,
+                        ),
+                    )
+
+                    repo = _repos.WorkflowRunRepo()
+
+                    # When
+                    outputs = repo.get_task_outputs(
+                        wf_run_id=run_id, task_inv_id=inv_id, config_name=config_name
+                    )
+
+                    # Then
+                    assert outputs == fake_outputs
 
             @staticmethod
             def test_invalid_inv_id(monkeypatch):
@@ -570,11 +646,11 @@ class TestWorkflowRunRepo:
                 inv_id = "inv1"
                 config_name = "<config sentinel>"
 
-                wf_run = Mock()
-                fake_outputs = ("<output sentinel 0>", "<output sentinel 1>")
-                wf_run.get_artifacts.return_value = {
-                    "non-existing-inv-id": fake_outputs,
-                }
+                task_run = create_autospec(sdk.TaskRun)
+                task_run.task_invocation_id = "<invalid inv ID>"
+
+                wf_run = create_autospec(sdk.WorkflowRun)
+                wf_run.get_tasks.return_value = {task_run}
 
                 by_id = Mock(return_value=wf_run)
                 monkeypatch.setattr(sdk.WorkflowRun, "by_id", by_id)