+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 🚨 *Breaking Changes*
+* Task results on QE have changed shape. This may cause some oddness when downloading older task artifacts.
 
 🔥 *Features*
 * New built-in config name - "auto" - used to submit workflows to a remote cluster when used inside Orquestra Studio.
@@ -15,10 +16,26 @@
 * Retry getting results from CE if the results were not ready but the workflow succeeded.
 * Using secrets inside the workflow function will now work correctly on Ray
 * Fix WorkflowDef.graph - honour kwargs of tasks and add aggregate_output to show outputs
+* Fixed returning intermediate workflow values (e.g. with `orq task results`) when the task has multiple outputs and only some of them were used in the rest of the workflow function. The following should work now as expected:
+```python
+@sdk.workflow
+def my_wf():
+    _, b = two_output_task()
+    all_outputs = two_output_task()
+    out1, out2 = all_outputs
+    return b, all_outputs, out1, out2
+```
+* Pickled workflow/task results should no longer cause workflows to fail inside the SDK machinery. Note: when passing a Python object between your tasks, you **must** ensure the Python dependencies are installed.
 
 💅 *Improvements*
+* `VersionMismatch` warning will be presented if we detect accessing a workflow def IR generated with another SDK version.
 
 🥷 *Internal*
+* `TaskOutputMetadata` model was added to the workflow def IR schema.
+* Workflows from CE return a new shape for workflow results and task results.
+* Workflows from Ray return a new shape for workflow and task results.
+* Workflows from QE return a new shape for task results.
+* Custom images will default to `None`, unless using Quantum Engine where the defaults stay the same.
 
 📃 *Docs*
 * _Resource Management_ guide should render properly now.
@@ -45,8 +62,8 @@ If a task defines its own imports (either source, dependencies, or both) - it wi
 
 🐛 *Bug Fixes*
 * Stopping a QE workflow after it has already stopped will no longer raise an exception.
-* Attempting to use the "in-process" runtime on the CLI will no longer raise an exception. Instead, a message teeling you to use the Python API or Ray will be printed.
 * Fix dependency issues causing CE workflows to fail if WF constant was library-dependent object.
+* Attempting to use the "in-process" runtime on the CLI will no longer raise an exception. Instead, a message telling you to use the Python API or Ray will be printed.
 
 🥷 *Internal*
 * During YAML conversion, Workflow SDK repo matched on host and path, not full URL.